nopbaitap

+++ module_2/src/exam/managerCodegym/model/Student.java
@@ -2,12 +2,12 @@ package exam.managerCodegym.model;
 
 public class Student extends Person{
     private String classes;
-    private float point;
+    private double point;
 
     public Student() {
     }
 
-    public Student(int id, String name, String dateOfBirth, boolean gender, String classes, float point) {
+    public Student(int id, String name, String dateOfBirth, boolean gender, String classes, double point) {
         super(id, name, dateOfBirth, gender);
         this.classes = classes;
         this.point = point;
@@ -21,7 +21,7 @@ public class Student extends Person{
         this.classes = classes;
     }
 
-    public float getPoint() {
+    public double getPoint() {
         return point;
     }
